Rename lib to libs and test and amend error response for get data
