+++ src/routes/data.route.ts
@@ -1,5 +1,5 @@
 import {Router, Request, Response} from 'express';
-import {parseReadingsPayload} from "../lib/readings-parser";
+import {parseReadingsPayload} from "../libs/readings-parser";
 import ReadingsRepository from "../repositories/readings.repository";
 
 export const dataRoute = Router();
@@ -15,23 +15,31 @@ dataRoute.post('/data', async (req: Request<{}, {}, string, {}, {}>, res) => {
   const parsedReadings = parseReadingsPayload(payload);
 
   if (parsedReadings.length === 0) {
-    res.status(400).json({success: false, error: "No valid data"}); // TODO! improve error reporting (add errors to parser response)
+    return res.status(400).json({success: false, error: "No valid data"}); // TODO! improve error reporting (add errors to parser response)
   }
 
   const dbRes = await readingsRepo.insertReadings(parsedReadings);
 
   if (dbRes.success) {
-    res.status(200).json({success: true});
+    return res.status(200).json({success: true});
   } else {
-
     return res.status(404).json({success: false});
   }
 });
 
-dataRoute.get('/data', async (req, res) => {
+dataRoute.get('/data', async (req: Request<{}, {}, {}, { to: string; from: string; }>, res) => {
   // TODO: check what dates have been requested, and retrieve all data within the given range
+  const queryParams = req.query;
+  const from = new Date(queryParams.from);
+  const to = new Date(queryParams.to);
+  if (isNaN(from.getDate()) || isNaN(to.getDate())) {
+    return res.status(400).json({
+      success: false,
+      error: "to and from query params need to be in ISO date format!!"
+    });
+  }
 
-  // getReading(...)
+  const readingsData = await readingsRepo.getReadings(from, to);
 
-  return res.json({success: false});
+  return res.json({success: false, data: readingsData});
 });
